+++ src/app/documentacion/controles-documentacion/controles-documentacion.component.html
@@ -5,14 +5,14 @@
 
 <div class="card border border-primary">
   <div class="card-body">
-    <app-com-find-tipo-documento-sunat 
+    <app-comp-tipo-documento-sunat 
       [filtros]="filtroro" 
       [SelectValorInicial]="documentoInit" 
       (onChange)="valorRecibido=$event"
       [mostrarSeries]=_mostrarSerie 
       [selectSerie]=null 
       [disabledSerie]=_disabledSerie>
-    </app-com-find-tipo-documento-sunat>
+    </app-comp-tipo-documento-sunat>
   </div>
 </div>
