+++ src/app/documentacion/documentacion.module.ts
@@ -5,64 +5,36 @@ import { DocumentacionRoutingModule } from './documentacion.routing';
 
 import { MainDocumentacionComponent } from './main-documentacion/main-documentacion.component';
 import { ControlesDocumentacionComponent } from './controles-documentacion/controles-documentacion.component';
-import { ComFindTipoDocumentoSunatModule } from '../intercambio/com-find-tipo-documento-sunat/com-find-tipo-documento-sunat.module';
-import { CompFindCodigoBarraComponent } from '../componentes/comp-find-codigo-barra/comp-find-codigo-barra.component';
 
-import { MatSlideToggleModule, MatSelectModule, MatAutocompleteModule } from '@angular/material';
+import { MatSlideToggleModule } from '@angular/material';
 import { MatTabsModule } from '@angular/material/tabs';
-import { CompFindAlmacenComponent } from '../componentes/comp-find-almacen/comp-find-almacen.component';
 import { ControlFindAlmacenComponent } from './control-find-almacen/control-find-almacen.component';
-import { MatInputModule } from '@angular/material/input';
 import { ControlFindCodigoBarraComponent } from './control-find-codigo-barra/control-find-codigo-barra.component';
-import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { ControlProveedorClienteComponent } from './control-proveedor-cliente/control-proveedor-cliente.component';
-import { CompTipoDocumentoIdentidadComponent } from '../componentes/comp-tipo-documento-identidad/comp-tipo-documento-identidad.component';
-import { CompFindProveedorClienteRucComponent } from '../componentes/comp-find-proveedor-cliente-ruc/comp-find-proveedor-cliente-ruc.component';
 import { ControlMedioPagoComponent } from './control-medio-pago/control-medio-pago.component';
-import { CompFindMedioPagoComponent } from '../componentes/comp-find-medio-pago/comp-find-medio-pago.component';
 import { ControlFindProductoComponent } from './control-find-producto/control-find-producto.component';
-import { CompFindProductoComponent } from '../componentes/comp-find-producto/comp-find-producto.component';
-import { PaginatorModule } from 'primeng/paginator';
-// import { CompFindProductoListModule } from '../componentes/comp-find-producto-list/comp-find-producto-list.module';
+import { ComponentesModule } from '../componentes/componentes.module';
 import { ControlFindMotivoIsAlmacenComponent } from './control-find-motivo-is-almacen/control-find-motivo-is-almacen.component';
-import { CompFindMotivoIsAlmacenComponent } from '../componentes/comp-find-motivo-is-almacen/comp-find-motivo-is-almacen.component';
-import { CompFindProveedorClienteListComponent } from '../componentes/comp-find-proveedor-cliente-list/comp-find-proveedor-cliente-list.component';
-import { MatIconModule } from '@angular/material/icon';
-import { CompFindProductoListComponent } from '../componentes/comp-find-producto-list/comp-find-producto-list.component';
 
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule, ReactiveFormsModule,
-    DocumentacionRoutingModule,    
-    PaginatorModule,
-    MatInputModule,
+    DocumentacionRoutingModule,
     MatSlideToggleModule,
     MatTabsModule,    
-    MatAutocompleteModule,
-    MatSelectModule,
-    MatProgressBarModule,
-    ComFindTipoDocumentoSunatModule,    
-    MatIconModule,
-    // CompFindProductoListModule
+    ComponentesModule
   ],
   declarations: [
     MainDocumentacionComponent, 
-    ControlesDocumentacionComponent,
-    CompFindAlmacenComponent,
+    ControlesDocumentacionComponent,    
     ControlFindAlmacenComponent,
-    CompFindCodigoBarraComponent,
-    CompTipoDocumentoIdentidadComponent,
-    CompFindProveedorClienteRucComponent,
-    CompFindMedioPagoComponent,
-    CompFindProductoComponent,
     ControlFindCodigoBarraComponent,
     ControlProveedorClienteComponent,    
-    ControlMedioPagoComponent, ControlFindProductoComponent, ControlFindMotivoIsAlmacenComponent,
-    CompFindMotivoIsAlmacenComponent,
-    CompFindProveedorClienteListComponent, 
-    CompFindProductoListComponent
+    ControlMedioPagoComponent, 
+    ControlFindProductoComponent,
+    ControlFindMotivoIsAlmacenComponent    
   ]
 })
